+++ src/main/java/com/example/ReviewZIP/ReviewZipApplication.java
@@ -2,10 +2,10 @@ package com.example.ReviewZIP;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
+import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
 
-@SpringBootApplication()
-public class ReviewZipApplication(exclude = SecurityAutoConfiguration.class) {
+@SpringBootApplication(exclude = SecurityAutoConfiguration.class)
+public class ReviewZipApplication {
 
 	public static void main(String[] args) {
 		SpringApplication.run(ReviewZipApplication.class, args);
