--- src/main/java/com/example/ReviewZIP/domain/hashtag/Hashtags.java
@@ -1,22 +1,0 @@
-package com.example.ReviewZIP.domain.hashtag;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Entity
-@Getter
-@Setter
-@NoArgsConstructor
-@Table(name = "hashtags")
-public class Hashtags {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id")
-    private Long id;
-
-    @Column(name = "name", length = 50)
-    private String name;
-
-}
