--- src/main/java/com/example/ReviewZIP/domain/hashtag/HashtagsController.java
@@ -1,4 +1,0 @@
-package com.example.ReviewZIP.domain.hashtag;
-
-public class HashtagsController {
-}
