--- src/main/java/com/example/ReviewZIP/domain/hashtag/HashtagsRepository.java
@@ -1,6 +1,0 @@
-package com.example.ReviewZIP.domain.hashtag;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface HashtagsRepository extends JpaRepository<Hashtags, Long> {
-}
