+++ src/main/java/com/example/ReviewZIP/domain/post/Posts.java
@@ -1,9 +1,10 @@
 package com.example.ReviewZIP.domain.post;
 
 import com.example.ReviewZIP.domain.image.Images;
-import com.example.ReviewZIP.domain.postHashtag.PostHashtags;
 import com.example.ReviewZIP.domain.postLike.PostLikes;
 import com.example.ReviewZIP.domain.user.Users;
+import com.example.ReviewZIP.global.entity.BaseEntity;
+import com.example.ReviewZIP.domain.postHashtag.PostHashtags;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
