+++ src/main/java/com/example/ReviewZIP/domain/postHashtag/PostHashtags.java
@@ -1,6 +1,5 @@
-package com.example.ReviewZIP.domain.posthashtag;
+package com.example.ReviewZIP.domain.postHashtag;
 
-import com.example.ReviewZIP.domain.hashtag.Hashtags;
 import com.example.ReviewZIP.domain.post.Posts;
 import jakarta.persistence.*;
 import lombok.Getter;
@@ -22,8 +21,5 @@ public class PostHashtags {
     @JoinColumn(name = "post_id")
     private Posts post;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "hashtag_id")
-    private Hashtags hashtag;
 
 }
