+++ src/main/java/com/example/ReviewZIP/domain/postHashtag/PostHashtagsRepository.java
@@ -1,4 +1,4 @@
-package com.example.ReviewZIP.domain.posthashtag;
+package com.example.ReviewZIP.domain.postHashtag;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PostHashtagsRepository extends JpaRepository<PostHashtags, Long> {
 
-    Page<PostHashtags> findByHashtagId(Long id, Pageable pageable);
+    Page<PostHashtags> findById(Long id, Pageable pageable);
 }
